# Avoid excessive allocation in Arena