--- a/src/Columns/ColumnString.h
+++ b/src/Columns/ColumnString.h
@@ -11,7 +11,6 @@
 #include <Common/memcmpSmall.h>
 #include <Common/assert_cast.h>
 #include <Core/Field.h>
-#include <Common/Arena.h>
 
 
 class Collator;
@@ -19,6 +18,8 @@ class Collator;
 
 namespace DB
 {
+
+class Arena;
 
 /** Column for String values.
   */