… tests/queries/0_stateless/02741_hashed_dictionary_load_factor.reference