### Changelog category (leave one):
- Performance Improvement


### Changelog entry (a user-readable short description of the changes that goes to CHANGELOG.md):
Queries with aggregation and a large number of streams will use less amount of memory during the plan's construction.