@alexey-milovidov 2023-11-23
Performance tests showed some uncertain change: 

https://s3.amazonaws.com/clickhouse-test-reports/57074/2ba2f480513a5b4561468c3d95bcd0c634936b44/performance_comparison_aarch64_[1_4]/report.html

I have validated it on my machine, just in case:

```
clickhouse-benchmark --query "SELECT sum(toNullable(toUInt8(number))) FROM numbers(100000000)"
```

But there is no difference:

```
New:

localhost:9000, queries: 140, QPS: 12.896, RPS: 1289696889.832, MiB/s: 9839.606, result RPS: 12.896, result MiB/s: 0.000.

0.000%          0.057 sec.
10.000%         0.059 sec.
20.000%         0.079 sec.
30.000%         0.079 sec.
40.000%         0.080 sec.
50.000%         0.080 sec.
60.000%         0.080 sec.
70.000%         0.080 sec.
80.000%         0.081 sec.
90.000%         0.082 sec.
95.000%         0.082 sec.
99.000%         0.083 sec.
99.900%         0.103 sec.
99.990%         0.103 sec.

Old:

localhost:9000, queries: 149, QPS: 12.708, RPS: 1270980883.133, MiB/s: 9696.815, result RPS: 12.708, result MiB/s: 0.000.

0.000%          0.057 sec.
10.000%         0.061 sec.
20.000%         0.077 sec.
30.000%         0.079 sec.
40.000%         0.079 sec.
50.000%         0.080 sec.
60.000%         0.080 sec.
70.000%         0.080 sec.
80.000%         0.081 sec.
90.000%         0.082 sec.
95.000%         0.082 sec.
99.000%         0.096 sec.
99.900%         0.203 sec.
99.990%         0.203 sec.
```